move agendaItem into domain
