+++ app/src/main/java/com/example/tasky/agenda_details/domain/repository/AgendaDetailsRemoteRepository.kt
@@ -1,9 +1,9 @@
 package com.example.tasky.agenda_details.domain.repository
 
+import com.example.tasky.agenda_details.data.model.EventResponse
+import com.example.tasky.agenda_details.data.model.ReminderResponse
+import com.example.tasky.agenda_details.data.model.TaskResponse
 import com.example.tasky.agenda_details.domain.model.EventDetails
-import com.example.tasky.agenda_details.domain.model.EventResponse
-import com.example.tasky.agenda_details.domain.model.ReminderResponse
-import com.example.tasky.agenda_details.domain.model.TaskResponse
 import com.example.tasky.common.domain.Result
 import com.example.tasky.common.domain.error.DataError
