+++ app/src/main/java/com/example/tasky/common/data/remote/TaskyApi.kt
@@ -1,8 +1,8 @@
 package com.example.tasky.common.data.remote
 
-import com.example.tasky.agenda_details.domain.model.EventResponse
-import com.example.tasky.agenda_details.domain.model.ReminderResponse
-import com.example.tasky.agenda_details.domain.model.TaskResponse
+import com.example.tasky.agenda_details.data.model.EventResponse
+import com.example.tasky.agenda_details.data.model.ReminderResponse
+import com.example.tasky.agenda_details.data.model.TaskResponse
 import com.example.tasky.common.domain.model.AccessTokenResponse
 import com.example.tasky.feature_agenda.domain.model.AgendaResponse
 import com.example.tasky.feature_agenda.domain.model.SyncAgendaResponse
