+++ app/src/main/java/com/example/tasky/feature_agenda/data/repository/AgendaRemoteRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.example.tasky.feature_agenda.data.repository
 
+import com.example.tasky.agenda_details.domain.model.AgendaItem
 import com.example.tasky.common.data.remote.TaskyApi
 import com.example.tasky.common.domain.Result
 import com.example.tasky.common.domain.error.DataError
@@ -7,7 +8,6 @@ import com.example.tasky.common.domain.util.toNetworkErrorType
 import com.example.tasky.di.AuthenticatedApi
 import com.example.tasky.feature_agenda.domain.model.SyncAgendaResponse
 import com.example.tasky.feature_agenda.domain.repository.AgendaRemoteRepository
-import com.example.tasky.feature_agenda.presentation.AgendaItem
 import java.io.IOException
 
 class AgendaRemoteRepositoryImpl(
