+++ app/src/main/java/com/example/tasky/feature_agenda/domain/model/AgendaResponse.kt
@@ -1,8 +1,8 @@
 package com.example.tasky.feature_agenda.domain.model
 
-import com.example.tasky.agenda_details.domain.model.EventResponse
-import com.example.tasky.agenda_details.domain.model.ReminderResponse
-import com.example.tasky.agenda_details.domain.model.TaskResponse
+import com.example.tasky.agenda_details.data.model.EventResponse
+import com.example.tasky.agenda_details.data.model.ReminderResponse
+import com.example.tasky.agenda_details.data.model.TaskResponse
 
 data class AgendaResponse(
     val events: List<EventResponse> = emptyList(),
